Move header related code to header module
